requirements.txt
pyautogui==0.9.54
opencv-python==4.13.0.92
pillow==12.2.0
pyyaml==6.0.3
pynput==1.8.1